// components/recruiting/types.ts
export interface RecruitingFormData {
  name: string;
  stream: string;
  course: string;
  year: string;
  experience: string;
  technologySkills: string[];
  learningTechnology: string[];
  resumeFile: File | null;
  photoFile: File | null;
  videoDriveLink: string;
}
</parameter>

<figma type="work">
Now I'll create separate form sections as components.